+++ R/dm_metabolic.R
@@ -10,7 +10,7 @@
 #' @keywords dataset metabolic
 #' @description A SDTM DM dataset for metabolic studies
 #' @docType data
-#' @format A data frame with 26 columns:
+#' @format A data frame with 28 columns:
 #'   \describe{
 #'     \item{STUDYID}{Study Identifier}
 #'     \item{DOMAIN}{Domain Abbreviation}
@@ -38,6 +38,8 @@
 #'     \item{COUNTRY}{Country}
 #'     \item{DMDTC}{Date/Time of Collection}
 #'     \item{DMDY}{Study Day of Collection}
+#'     \item{ARMNRS}{Reason Arm and/or Actual Arm is Null}
+#'     \item{ACTARMUD}{Description of Unplanned Actual Arm}
 #'   }
 #'
 #' @source Constructed by {admiralmetabolic} developers