+++ R/dm_neuro.R
@@ -10,7 +10,7 @@
 #' @keywords dataset neurology
 #' @description A SDTM DM dataset for Alzheimer's disease studies
 #' @docType data
-#' @format A data frame with 27 columns:
+#' @format A data frame with 28 columns:
 #'   \describe{
 #'     \item{STUDYID}{Study Identifier}
 #'     \item{DOMAIN}{Domain Abbreviation}
@@ -25,8 +25,8 @@
 #'     \item{DTHDTC}{Date/Time of Death}
 #'     \item{DTHFL}{Subject Death Flag}
 #'     \item{SITEID}{Study Site Identifier}
-#'     \item{AGE}{Age}
 #'     \item{BRTHDTC}{Date/Time of Birth}
+#'     \item{AGE}{Age}
 #'     \item{AGEU}{Age Units}
 #'     \item{SEX}{Sex}
 #'     \item{RACE}{Race}
@@ -39,6 +39,7 @@
 #'     \item{DMDTC}{Date/Time of Collection}
 #'     \item{DMDY}{Study Day of Collection}
 #'     \item{ARMNRS}{Reason Arm and/or Actual Arm is Null}
+#'     \item{ACTARMUD}{Description of Unplanned Actual Arm}
 #'   }
 #'
 #' @source Constructed by {admiralneuro} developers