+++ R/dm_peds.R
@@ -10,7 +10,7 @@
 #' @keywords dataset pediatrics
 #' @description An updated SDTM DM dataset with pediatric patients
 #' @docType data
-#' @format A data frame with 26 columns:
+#' @format A data frame with 28 columns:
 #'   \describe{
 #'     \item{STUDYID}{Study Identifier}
 #'     \item{DOMAIN}{Domain Abbreviation}
@@ -25,6 +25,7 @@
 #'     \item{DTHDTC}{Date/Time of Death}
 #'     \item{DTHFL}{Subject Death Flag}
 #'     \item{SITEID}{Study Site Identifier}
+#'     \item{BRTHDTC}{Date/Time of Birth}
 #'     \item{AGE}{Age}
 #'     \item{AGEU}{Age Units}
 #'     \item{SEX}{Sex}
@@ -37,7 +38,8 @@
 #'     \item{COUNTRY}{Country}
 #'     \item{DMDTC}{Date/Time of Collection}
 #'     \item{DMDY}{Study Day of Collection}
-#'     \item{BRTHDTC}{Date/Time of Birth}
+#'     \item{ARMNRS}{Reason Arm and/or Actual Arm is Null}
+#'     \item{ACTARMUD}{Description of Unplanned Actual Arm}
 #'   }
 #'
 #' @source Constructed by {admiralpeds} developers